Move logic to firewall deployment controller.
Cannot be done on set-level because the sets will toggle each other's status.
